Fixed names and summaries.
