+++ UserInterface/UI.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AnimalsTypes\AnimalsTypes.csproj" />
+    <ProjectReference Include="..\AnimalTypes\AnimalTypes.csproj" />
     <ProjectReference Include="..\ConstantsRepository\ConstantsRepository.csproj" />
     <ProjectReference Include="..\GameEngine\GameEngine.csproj" />
   </ItemGroup>
